requirements.txt
langchain-gigachat
langchain-core
langchain-mcp-adapters
langgraph
python-dotenv
docxtpl
context7-mcp 